preprocessing/requirements_preprocessing.txt
fastapi
uvicorn[standard]
python-dotenv
numpy
open3d==0.18.0
pydantic
matplotlib